pyproject.toml
[project]
name = "fprofourteen"
version = "0.1.0"
description = "C言語の成果物を確認するためのPythonスクリプト"
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["pillow>=12.1.0"]
